# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Our Empire Satta Matka API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export interface RegisterInput {
  fullName: string;
  phone: string;
  /** @nullable */
  email?: string | null;
  password: string;
  /** @nullable */
  referralCode?: string | null;
}

export interface LoginInput {
  phone: string;
  password: string;
}

export interface RefreshInput {
  refreshToken: string;
}

export interface ForgotPasswordInput {
  phone: string;
}

export interface ResetPasswordInput {
  phone: string;
  otp: string;
  newPassword: string;
}

export interface User {
  id: number;
  fullName: string;
  phone: string;
  /** @nullable */
  email?: string | null;
  balance: string;
  referralCode?: string;
  status: string;
  createdAt: string;
}

export interface AuthResponse {
  accessToken: string;
  refreshToken: string;
  user: User;
}

export interface TokenResponse {
  accessToken: string;
}

export interface AdminLoginInput {
  username: string;
  password: string;
}

export interface AdminUser {
  id: number;
  username: string;
  role: string;
}

export interface AdminAuthResponse {
  accessToken: string;
  admin: AdminUser;
}

export interface UserUpdate {
  fullName?: string;
  /** @nullable */
  email?: string | null;
}

export interface UserDetail {
  id: number;
  fullName: string;
  phone: string;
  /** @nullable */
  email?: string | null;
  balance: string;
  referralCode?: string;
  /** @nullable */
  referredBy?: number | null;
  status: string;
  createdAt: string;
  totalBets?: number;
  totalDeposits?: string;
  totalWithdrawals?: string;
}

export interface BalanceResponse {
  balance: string;
}

export interface ChangePasswordInput {
  currentPassword: string;
  newPassword: string;
}

export interface Referral {
  id: number;
  fullName: string;
  phone: string;
  createdAt: string;
  earnings: string;
}

export interface UserAnalytics {
  totalBets: number;
  totalWagered: string;
  totalWon: string;
  totalLost: string;
  winRate: number;
  /** @nullable */
  favoriteMarket?: string | null;
  profitLoss?: string;
}

export interface Notification {
  id: number;
  title: string;
  message: string;
  isRead: boolean;
  createdAt: string;
}

export interface Result {
  id: number;
  marketId: number;
  marketName?: string;
  date: string;
  /** @nullable */
  openNumber?: string | null;
  /** @nullable */
  closeNumber?: string | null;
  /** @nullable */
  jodiNumber?: string | null;
  status: string;
}

export interface Market {
  id: number;
  name: string;
  openTime: string;
  closeTime: string;
  status: string;
  minBet: string;
  maxBet: string;
  isBettingOpen?: boolean;
  todayResult?: Result | null;
}

export interface MarketInput {
  name: string;
  openTime: string;
  closeTime: string;
  minBet?: string;
  maxBet?: string;
  status?: string;
}

export interface MarketUpdate {
  name?: string;
  openTime?: string;
  closeTime?: string;
  minBet?: string;
  maxBet?: string;
  status?: string;
}

export type BetInputGameType = typeof BetInputGameType[keyof typeof BetInputGameType];


export const BetInputGameType = {
  jantri: 'jantri',
  open: 'open',
  crossing: 'crossing',
  no_to_no: 'no_to_no',
} as const;

export interface BetNumber {
  number: string;
  amount: string;
}

export interface BetInput {
  marketId: number;
  gameType: BetInputGameType;
  numbers: BetNumber[];
  totalAmount: string;
}

export interface Bet {
  id: number;
  userId?: number;
  marketId: number;
  marketName?: string;
  gameType: string;
  numbers: BetNumber[];
  totalAmount: string;
  status: string;
  winAmount?: string;
  createdAt: string;
}

export interface WinningsResponse {
  totalWinnings: string;
  bets: Bet[];
}

export interface ResultInput {
  marketId: number;
  date: string;
  /** @nullable */
  openNumber?: string | null;
  /** @nullable */
  closeNumber?: string | null;
  /** @nullable */
  jodiNumber?: string | null;
  status?: string;
}

export interface ResultUpdate {
  /** @nullable */
  openNumber?: string | null;
  /** @nullable */
  closeNumber?: string | null;
  /** @nullable */
  jodiNumber?: string | null;
  status?: string;
}

export interface Transaction {
  id: number;
  userId: number;
  type: string;
  amount: string;
  status: string;
  /** @nullable */
  method?: string | null;
  /** @nullable */
  utrNumber?: string | null;
  /** @nullable */
  adminNote?: string | null;
  createdAt: string;
  /** @nullable */
  userName?: string | null;
}

export interface AddMoneyInput {
  amount: string;
  utrNumber: string;
  method?: string;
}

export interface WithdrawalInput {
  amount: string;
  method?: string;
  /** @nullable */
  upiId?: string | null;
  /** @nullable */
  accountNumber?: string | null;
  /** @nullable */
  ifscCode?: string | null;
  /** @nullable */
  accountHolderName?: string | null;
}

export interface UpiInfo {
  upiId: string;
  /** @nullable */
  qrCodeUrl?: string | null;
}

export interface AdminDashboard {
  totalUsers: number;
  newUsersToday?: number;
  totalBetsToday: number;
  totalAmountToday: string;
  pendingDeposits: number;
  pendingWithdrawals: number;
  activeMarkets: number;
  totalRevenue?: string;
}

export interface AdminUsersResponse {
  users: UserDetail[];
  total: number;
}

export interface UserStatusUpdate {
  status: string;
}

export interface BalanceAdjust {
  amount: string;
  reason: string;
}

export interface AdminTransactionsResponse {
  transactions: Transaction[];
  total: number;
}

export interface AdminBetsResponse {
  bets: Bet[];
  total: number;
}

export interface NumberStat {
  number: string;
  betCount: number;
  totalAmount: string;
}

export interface BetAnalytics {
  totalBets: number;
  totalAmount: string;
  numberStats: NumberStat[];
}

export interface AppSettings {
  appName: string;
  whatsappNumber: string;
  telegramLink: string;
  maintenanceMode?: boolean;
  minDeposit?: string;
  minWithdrawal?: string;
  referralCommission?: string;
  upiIds?: string[];
}

export interface AppSettingsUpdate {
  appName?: string;
  whatsappNumber?: string;
  telegramLink?: string;
  maintenanceMode?: boolean;
  minDeposit?: string;
  minWithdrawal?: string;
  referralCommission?: string;
  upiIds?: string[];
}

export interface NotificationInput {
  title: string;
  message: string;
  targetType?: string;
  targetUserIds?: number[];
}

export interface RejectInput {
  reason: string;
}

export interface RevenuePeriod {
  date: string;
  revenue: string;
  deposits: string;
  withdrawals: string;
}

export interface RevenueAnalytics {
  period: string;
  totalRevenue: string;
  totalDeposits?: string;
  totalWithdrawals?: string;
  data: RevenuePeriod[];
}

export type GetMyBetsParams = {
status?: string;
marketId?: number;
limit?: number;
offset?: number;
};

export type GetTransactionsParams = {
type?: string;
limit?: number;
offset?: number;
};

export type AdminGetUsersParams = {
search?: string;
status?: string;
limit?: number;
offset?: number;
};

export type AdminGetDepositsParams = {
status?: string;
limit?: number;
offset?: number;
};

export type AdminGetWithdrawalsParams = {
status?: string;
limit?: number;
offset?: number;
};

export type AdminGetBetsParams = {
marketId?: number;
status?: string;
userId?: number;
limit?: number;
offset?: number;
};

export type AdminGetBetAnalyticsParams = {
marketId?: number;
date?: string;
};

export type AdminGetRevenueAnalyticsParams = {
period?: string;
};