# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Our Empire Satta Matka API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: registerUser
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: loginUser
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/refresh:
    post:
      operationId: refreshToken
      tags: [auth]
      summary: Refresh access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
  /auth/forgot-password:
    post:
      operationId: forgotPassword
      tags: [auth]
      summary: Request password reset OTP
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForgotPasswordInput"
      responses:
        "200":
          description: OTP sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /auth/reset-password:
    post:
      operationId: resetPassword
      tags: [auth]
      summary: Reset password with OTP
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordInput"
      responses:
        "200":
          description: Password reset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # USER
...
  /markets/{id}:
    get:
      operationId: getMarket
      tags: [markets]
      summary: Get market by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Market"
  /markets/{id}/results:
    get:
      operationId: listMarketResults
      tags: [markets]
      summary: Get market results
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Result"

  # BETS
  /bets/place:
    post:
      operationId: placeBet
      tags: [bets]
      summary: Place a bet
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BetInput"
      responses:
        "201":
          description: Bet placed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bet"
        "400":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /bets/my-bets:
    get:
      operationId: getMyBets
      tags: [bets]
      summary: Get user bets
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: marketId
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Bet"
  /bets/winnings:
    get:
      operationId: getWinnings
      tags: [bets]
      summary: Get user winnings
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WinningsResponse"
  /bets/{id}/cancel:
    post:
      operationId: cancelBet
      tags: [bets]
      summary: Cancel a pending bet
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # WALLET
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
    RegisterInput:
      type: object
      required: [fullName, phone, password]
      properties:
        fullName:
          type: string
        phone:
          type: string
        email:
          type: ["string", "null"]
        password:
          type: string
        referralCode:
          type: ["string", "null"]
    LoginInput:
      type: object
      required: [phone, password]
      properties:
        phone:
          type: string
        password:
          type: string
    RefreshInput:
      type: object
      required: [refreshToken]
      properties:
        refreshToken:
          type: string
    ForgotPasswordInput:
      type: object
      required: [phone]
      properties:
        phone:
          type: string
    ResetPasswordInput:
      type: object
      required: [phone, otp, newPassword]
      properties:
        phone:
          type: string
        otp:
          type: string
        newPassword:
          type: string
    AuthResponse:
      type: object
      required: [accessToken, refreshToken, user]
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        user:
          $ref: "#/components/schemas/User"
    TokenResponse:
      type: object
      required: [accessToken]
      properties:
        accessToken:
          type: string
...
    User:
      type: object
      required: [id, fullName, phone, balance, status, createdAt]
      properties:
        id:
          type: integer
        fullName:
          type: string
        phone:
          type: string
        email:
          type: ["string", "null"]
        balance:
          type: string
        referralCode:
          type: string
        status:
          type: string
        createdAt:
          type: string
...
    Market:
      type: object
      required: [id, name, openTime, closeTime, status, minBet, maxBet]
      properties:
        id:
          type: integer
        name:
          type: string
        openTime:
          type: string
        closeTime:
          type: string
        status:
          type: string
        minBet:
          type: string
        maxBet:
          type: string
        isBettingOpen:
          type: boolean
        todayResult:
          $ref: "#/components/schemas/Result"
          nullable: true
...
    BetInput:
      type: object
      required: [marketId, gameType, numbers, totalAmount]
      properties:
        marketId:
          type: integer
        gameType:
          type: string
          enum: [jantri, open, crossing, no_to_no]
        numbers:
          type: array
          items:
            $ref: "#/components/schemas/BetNumber"
        totalAmount:
          type: string
    BetNumber:
      type: object
      required: [number, amount]
      properties:
        number:
          type: string
        amount:
          type: string
    Bet:
      type: object
      required: [id, marketId, gameType, numbers, totalAmount, status, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        marketId:
          type: integer
        marketName:
          type: string
        gameType:
          type: string
        numbers:
          type: array
          items:
            $ref: "#/components/schemas/BetNumber"
        totalAmount:
          type: string
        status:
          type: string
        winAmount:
          type: string
        createdAt:
          type: string
    WinningsResponse:
      type: object
      required: [totalWinnings, bets]
      properties:
        totalWinnings:
          type: string
        bets:
          type: array
          items:
            $ref: "#/components/schemas/Bet"
    Result:
      type: object
      required: [id, marketId, date, status]
      properties:
        id:
          type: integer
        marketId:
          type: integer
        marketName:
          type: string
        date:
          type: string
        openNumber:
          type: ["string", "null"]
        closeNumber:
          type: ["string", "null"]
        jodiNumber:
          type: ["string", "null"]
        status:
          type: string
...