# lib/api-zod/src/generated/types/market.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Our Empire Satta Matka API
 * OpenAPI spec version: 0.1.0
 */
import type { Result } from './result';

export interface Market {
  id: number;
  name: string;
  openTime: string;
  closeTime: string;
  status: string;
  minBet: string;
  maxBet: string;
  isBettingOpen?: boolean;
  todayResult?: Result | null;
}