# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Our Empire Satta Matka API
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  id: number;
  fullName: string;
  phone: string;
  /** @nullable */
  email?: string | null;
  balance: string;
  referralCode?: string;
  status: string;
  createdAt: string;
}