# lib/api-zod/src/generated/types/winningsResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Our Empire Satta Matka API
 * OpenAPI spec version: 0.1.0
 */
import type { Bet } from './bet';

export interface WinningsResponse {
  totalWinnings: string;
  bets: Bet[];
}